+++ data/media/galerie-bannieres.js
@@ -1,0 +1,6 @@
+(function () {
+  var script = document.createElement("script");
+  script.src = "../data/gallery/bannieres.js";
+  script.async = false;
+  document.currentScript.after(script);
+}());